+++ e2e/env/env.go
@@ -647,5 +647,6 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"oci environment singularityenv": c.ociSingularityEnv,
 		"oci environment option":         c.ociEnvOption,
 		"oci environment file":           c.ociEnvFile,
+		"oci native env eval":            c.ociNativeEnvEval,
 	}
 }